[Feature]: Shared variable done!
